refactor: remove MailboxTestUtil.java in favor of MailboxSetupHelper
- allows keeping track of state and doing a proper cleanup of the temporary folder
