+++ store/src/test/java/com/zextras/mailbox/MailboxTestSuite.java
@@ -8,6 +8,7 @@ package com.zextras.mailbox;
 
 import com.zextras.mailbox.util.MailboxTestData;
 import com.zextras.mailbox.util.MailboxTestExtension;
+import com.zextras.mailbox.util.MailboxSetupHelper;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
 public abstract class MailboxTestSuite {
@@ -17,6 +18,7 @@ public abstract class MailboxTestSuite {
 			"f4806430-b434-4e93-9357-a02d9dd796b8");
 
 	@RegisterExtension
-	protected static MailboxTestExtension mailboxTestExtension = new MailboxTestExtension(mailboxTestData);
+	protected static MailboxTestExtension mailboxTestExtension = new MailboxTestExtension(mailboxTestData,
+			MailboxSetupHelper.create());
 
 }
